--- ACPCore/Assets/Editor/BuildPostProcessor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 46ffd0e8adc6049e58aac1095248fe06
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
