+++ ACPCore/Assets/Plugins/iOS/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3b339f6b7b18840ce880b492124b92f4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
