+++ README.md
@@ -14,7 +14,7 @@
 
 ## Prerequisites
 
-The `Unity Hub` application is required for development and testing. Inside of `Unity Hub`, you will be required to download the current version of the `Unity` app.
+The `Unity Hub` application is required for development and testing. Inside of `Unity Hub`, you will be required to download the `Unity` app. The ACPCore Unity package is built using Unity version 2019.4.
 
 [Download the Unity Hub](http://unity3d.com/unity/download). The free version works for development and testing, but a Unity Pro license is required for distribution. See [Distribution](#distribution) below for details.
 
@@ -27,7 +27,17 @@ Plugins for a Unity project use the following folder structure:
 - Download [ACPCore-1.0.0-Unity.zip](./bin/ACPCore-1.0.0-Unity.zip) 
 - Unzip `ACPCore-1.0.0-Unity.zip`
 - Import `ACPCore.unitypackage` via Assets-Import Package
-- 
+
+#### Android
+No additional steps are required for Android installation.
+
+#### iOS
+ACPCore 1.0.0 and above is shipped with XCFrameworks. Follow these steps to add them to the Xcode project generated when building and running for iOS platform in Unity.
+1. Go to File -> Project Settings -> Build System and select `New Build System`.
+2. [Download](https://github.com/Adobe-Marketing-Cloud/acp-sdks/tree/master/iOS/ACPCore) `ACPCore.xcframework`, `ACPIdentity.xcframework`, `ACPLifecycle.xcframework` and `ACPSignal.xcframework`.
+3. Select the UnityFramework target -> Go to Build Phases tab -> Add the XCFrameworks downloaded in Step 2 to `Link Binary with Libraries`.
+4. Select the Unity-iPhone target -> Go to Build Phases tab -> Add the XCFrameworks downloaded in Steps 2 to `Link Binary with Libraries` and `Embed Frameworks`. Alternatively, select `Unity-iPhone` target -> Go to `General` tab -> Add the XCFrameworks downloaded in Steps 2 to `Frameworks, Libraries, and Embedded Content` -> Select `Embed and sign` option.
+
 ## Usage
 
 ### [Core](https://aep-sdks.gitbook.io/docs/using-mobile-extensions/mobile-core)
@@ -304,28 +314,30 @@ To build demo app for specific platform follow the below instructions.
 
 ###### Android
 1. Make sure you have an Android device connected.
-1. From the menu of the `Unity` app, select __File > Build Settings...__
-1. Select `Android` from the __Platform__ window
-1. If `Android` is not the active platform, hit the button that says __Switch Platform__ (it will only be available if you actually need to switch active platforms)
-1. Press the __Build And Run__ button
-1. You will be asked to provide a location to save the build. Make a new directory at *unity-acpcore/ACPCore/Builds* (this folder is in the .gitignore file)
-1. Name build whatever you want and press __Save__
-1. `Unity` will build an `apk` file and automatically deploy it to the connected device
+2. From the menu of the `Unity` app, select __File > Build Settings...__
+3. Select `Android` from the __Platform__ window
+4. If `Android` is not the active platform, hit the button that says __Switch Platform__ (it will only be available if you actually need to switch active platforms)
+5. Press the __Build And Run__ button
+6. You will be asked to provide a location to save the build. Make a new directory at *unity-acpcore/ACPCore/Builds* (this folder is in the .gitignore file)
+7. Name build whatever you want and press __Save__
+8. `Unity` will build an `apk` file and automatically deploy it to the connected device
 
 ###### iOS
 1. From the menu of the `Unity` app, select __File > Build Settings...__
-1. Select `iOS` from the __Platform__ window
-1. If `iOS` is not the active platform, hit the button that says __Switch Platform__ (it will only be available if you actually need to switch active platforms)
-1. Press the __Build And Run__ button
-1. You will be asked to provide a location to save the build. Make a new directory at *unity-acpcore/ACPCore/Builds* (this folder is in the .gitignore file)
-1. Name build whatever you want and press __Save__
-1. `Unity` will create and open an `Xcode` project
-1. From the Xcode project run the app on a simulator.
-1. If you get an error `Symbol not found: _OBJC_CLASS_$_WKWebView`. Select the Unity-iPhone target -> Go to Build Phases tab -> Add `Webkit.Framework` to `Link Binary with Libraries`.
+2. Select `iOS` from the __Platform__ window
+3. If `iOS` is not the active platform, hit the button that says __Switch Platform__ (it will only be available if you actually need to switch active platforms)
+4. Press the __Build And Run__ button
+5. You will be asked to provide a location to save the build. Make a new directory at *unity-acpcore/ACPCore/Builds* (this folder is in the .gitignore file)
+6. Name build whatever you want and press __Save__
+7. `Unity` will create and open an `Xcode` project
+8. From the Xcode project run the app on a simulator.
+9. [Download](https://github.com/Adobe-Marketing-Cloud/acp-sdks/tree/master/iOS/ACPCore) `ACPCore.xcframework`, `ACPIdentity.xcframework`, `ACPLifecycle.xcframework` and `ACPSignal.xcframework`.
+10. Select the UnityFramework target -> Go to Build Phases tab -> Add the XCFrameworks downloaded in Step 9 to `Link Binary with Libraries`.
+11. Select the Unity-iPhone target -> Go to Build Phases tab -> Add the XCFrameworks downloaded in Steps 9 to `Link Binary with Libraries` and `Embed Frameworks`. Alternatively, select `Unity-iPhone` target -> Go to `General` tab -> Add the XCFrameworks downloaded in Steps 9 to `Frameworks, Libraries, and Embedded Content` -> Select `Embed and sign` option.
 
-## Additional Cordova Plugins
+## Additional Unity Plugins
 
-Below is a list of additional Unity plugins from the AEP SDK suite:
+Below is a list of additional Unity plugins from the ACP SDK suite:
 
 | Extension | GitHub | Unity Package |
 |-----------|--------|-----|
